Removed duplicate configuration by leaving backend as string a little longer
